+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Unreleased
 
 * Added a program for plotting magnetic field lines
+* Added an option to footpoint() and farpoint() functions to normalise the output coordinates to the light cylinder radius.
 
 ----
 ## v1.1.1 (2018-03-06)